Fixed segfault from gROOT.SetBatch with argparse
I'm not entirely sure how it happens, but the gROOT.SetBatch() function
results in a segfault coming from the cmssw /lib/ROOT.py file when used in tandem with the argparse module.
Adding "-b-" to argv is an equivalent way to enter root batch mode that
does not cause this error.